refactor: create-drip-list button opens as modal
