+++ src/lib/components/slots/wallet-slot.svelte
