+++ src/lib/flows/create-drip-list-flow/steps/configure-one-time-donation/configure-one-time-donation.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
   import Button from '$lib/components/button/button.svelte';
   import Check from 'radicle-design-system/icons/Check.svelte';
-  import StandaloneFlowStepLayout from '../../../components/standalone-flow-step-layout/standalone-flow-step-layout.svelte';
+  import StandaloneFlowStepLayout from '$lib/components/standalone-flow-step-layout/standalone-flow-step-layout.svelte';
   import { createEventDispatcher } from 'svelte';
   import type { StepComponentEvents } from '$lib/components/stepper/types';
   import ArrowLeftIcon from 'radicle-design-system/icons/ArrowLeft.svelte';
   import type { Writable } from 'svelte/store';
-  import type { State } from '../../funder-onboarding-flow';
+  import type { State } from '../../create-drip-list-flow';
   import OneTimeDonationEditor from '$lib/components/one-time-donation-editor/one-time-donation-editor.svelte';
 
   const dispatch = createEventDispatcher<StepComponentEvents>();
