+++ src/lib/flows/create-drip-list-flow/steps/connect-wallet/connect-wallet.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import AccountBox from '$lib/components/account-box/account-box.svelte';
   import FormField from '$lib/components/form-field/form-field.svelte';
-  import StandaloneFlowStepLayout from '../../../components/standalone-flow-step-layout/standalone-flow-step-layout.svelte';
+  import StandaloneFlowStepLayout from '$lib/components/standalone-flow-step-layout/standalone-flow-step-layout.svelte';
   import walletStore from '$lib/stores/wallet/wallet.store';
   import ArrowRightIcon from 'radicle-design-system/icons/ArrowRight.svelte';
   import ArrowLeftIcon from 'radicle-design-system/icons/ArrowLeft.svelte';
